Adding different entity types
Moving most content of old sensor.py to new file entity.py, leaving behind sensor specific stuff in sensor.py and adding a new switch.py with switch specific stuff.

Moving entity setup to const.py
